requirements.txt
google-cloud-pubsub >= 0.39.1
google-cloud-storage >= 1.13.2